finished encoder decoder

+++ Server/src/main/java/bgu/spl/net/impl/BGRSServer/Commands/Command.java
@@ -5,8 +5,8 @@ import bgu.spl.net.impl.BGRSServer.Message;
 
 public abstract class Command {
     protected User user;
-    protected final int ACK_OP_CODE = 12;
-    protected final int ERROR_OP_CODE = 13;
+    protected final short ACK_OP_CODE = 12;
+    protected final short ERROR_OP_CODE = 13;
 
     public abstract Message execute();
     public User getUser () { return user; }
